tests cleanup & address feedback
