+++ posthog/api/test/test_user.py
@@ -8,70 +8,7 @@ from posthog.test.base import APITransactionBaseTest, BaseTest
 
 
 class TestUser(BaseTest):
-    TESTS_API = True  # TODO: Remove once endpoint is refactored to DRF
-
-    def test_create_user_with_distinct_id(self):
-        with self.settings(TEST=False):
-            user = User.objects.create_user(first_name="Tim", email="tim@gmail.com", password=None)
-        self.assertNotEqual(user.distinct_id, "")
-        self.assertNotEqual(user.distinct_id, None)
-
-    def test_analytics_metadata(self):
-
-        # One org, one team, anonymized
-        organization, team, user = User.objects.bootstrap(
-            company_name="Test Org", email="test_org@posthog.com", password="12345678", anonymize_data=True,
-        )
-
-        with self.settings(EE_AVAILABLE=True, MULTI_TENANCY=True):
-            self.assertEqual(
-                user.get_analytics_metadata(),
-                {
-                    "realm": "cloud",
-                    "ee_available": True,
-                    "email_opt_in": False,
-                    "anonymize_data": True,
-                    "email": None,
-                    "is_signed_up": True,
-                    "organization_count": 1,
-                    "project_count": 1,
-                    "team_member_count_all": 1,
-                    "completed_onboarding_once": False,
-                    "billing_plan": None,
-                    "organization_id": str(organization.id),
-                    "project_id": str(team.uuid),
-                    "project_setup_complete": False,
-                },
-            )
-
-        # Multiple teams, multiple members, completed onboarding
-        user = User.objects.create(email="test_org_2@posthog.com", email_opt_in=True)
-        OrganizationMembership.objects.create(user=user, organization=self.organization)
-        Team.objects.create(organization=self.organization)
-        self.team.completed_snippet_onboarding = True
-        self.team.ingested_event = True
-        self.team.save()
-
-        with self.settings(EE_AVAILABLE=False, MULTI_TENANCY=False):
-            self.assertEqual(
-                user.get_analytics_metadata(),
-                {
-                    "realm": "hosted",
-                    "ee_available": False,
-                    "email_opt_in": True,
-                    "anonymize_data": False,
-                    "email": "test_org_2@posthog.com",
-                    "is_signed_up": True,
-                    "organization_count": 1,
-                    "project_count": 2,
-                    "team_member_count_all": 2,
-                    "completed_onboarding_once": True,
-                    "billing_plan": None,
-                    "organization_id": str(self.organization.id),
-                    "project_id": str(self.team.uuid),
-                    "project_setup_complete": True,
-                },
-            )
+    TESTS_API = True
 
     # TODO: Move to TestUserAPI once endpoint is refactored to DRF
     def test_user_team_update(self):
@@ -181,7 +118,7 @@ class TestUserAPI(APITransactionBaseTest):
             self.user.distinct_id,
             {
                 "realm": "hosted",
-                "ee_available": True,
+                "is_ee_available": True,
                 "email_opt_in": False,
                 "anonymize_data": False,
                 "email": "user1@posthog.com",
